requirements.txt
agents==1.4.0
openai==1.70.0
python-dotenv==1.1.0
